Remove problematic color profile from CTU logo files
Currently, Zero eLab Viewer prints the following warning message during
startup:

> libpng warning: iCCP: known incorrect sRGB profile

The CTU logo PNGs contain the IEC61966-2.1 color profile. This profile
reportedly can cause problems [1] and is non-free [2].

To resolve the issue and stop the warning, this commit alters the images
to remove the color profile. The images look the same in GIMP.
The files were updated using the following command [3]:

```
$ convert image.png +profile "icc" image.png
```

[1]: https://github.com/pnggroup/libpng/blob/68d7ce8040a611e9166890e02ecc409a03d76c2e/png.c#L2372
[2]: https://github.com/google/cadvisor/issues/915
[3]: https://github.com/nodejs/node/pull/5811