fix: Re-bake the Windows ICO with the minimum required bitmaps
For some not-entirely-known reason Windows Explorer on Windows 11
refuses to show the embedded 48x48 EXE icon on the desktop with the
current icon file. Instead, it shows something that looks like an
upscaled lower-resolution icon.

Dropping the 64x64 and 128x128 icons seemed to fix the issue. I think
this might be because 128x128 is not a icon supported for EXE files.

The size set comes from a recommendation at
https://learn.microsoft.com/en-us/windows/apps/design/style/iconography/app-icon-construction#icon-sizes-win32

> Apps should have, at the bare minimum: 16x16, 24x24, 32x32, 48x48,
> and 256x256. This covers the most common icon sizes, and by providing
> a 256px icon, ensures Windows should only ever scale your icon down,
> never up.